backend/app/config.py
"""
Configuration file for the backend
"""
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Telegram configuration
TELEGRAM_BOT_TOKEN = os.getenv('TELEGRAM_BOT_TOKEN')
TELEGRAM_BOT_USERNAME = os.getenv("TELEGRAM_BOT_USERNAME")

# Auth settings
JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key")
JWT_ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Redis settings
REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379/0")
VERIFICATION_CODE_TTL = 300  # 5 minutes in seconds 