requirements.txt
requests==2.31.0
python-dotenv==1.0.0
telethon==1.34.0
jinja2==3.1.3
markdown2==2.4.12
aiohttp==3.9.3
pytz==2024.1
openai==1.13.3

fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==5.0.3

python-telegram-bot[job-queue]==20.8
PyJWT==2.8.0